% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{archive_source_resource}
\alias{archive_source_resource}
\title{Package Archive Source Code Resource Class}
\usage{
archive_source_resource(path = NA_character_)
}
\description{
The extracted source code from a package's build archive.
}
