% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pkg_data_derive.R
\name{capture_pkg_data_derive}
\alias{capture_pkg_data_derive}
\title{Derive data and capture output}
\usage{
capture_pkg_data_derive(pkg, resource, field, ..., quiet = opt("quiet"))
}
\arguments{
\item{pkg}{A \code{\link[=pkg]{pkg()}}}

\item{resource}{A \code{\link[=resource]{resource()}}, or if not provided, the \code{\link{resource}}
extracted from \code{pkg@resource}.}

\item{field}{Used for dispatching on which field to derive. Methods are
provided such that a simple \code{character} field name can be passed and
used to build a class for dispatching to the right derivation function.}

\item{...}{Used by specific methods.}
}
\description{
Uses \code{evaluate::evaluate} to capture execution logs.
}
\keyword{internal}
