% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metric_data_frame.R
\name{class_package_matrix}
\alias{class_package_matrix}
\title{A matrix produced by \code{available.packages}}
\usage{
class_package_matrix(.data = character(0))
}
\arguments{
\item{.data}{See \code{\link[S7:class_character]{S7::class_character}}}
}
\description{
A matrix produced by \code{available.packages}
}
\note{
Due to a \href{https://github.com/RConsortium/S7/issues/401}{limitation of \code{S7}'s handling of \code{matrix} and \code{array} classes}, we currently need
to wrap matrices that we want to handle through \code{convert} in our own custom
classes. This may not be necessary in the future.
}
