% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_evaluate.R
\name{cli_singleton_html_style}
\alias{cli_singleton_html_style}
\title{Produce a html style for cli output}
\usage{
cli_singleton_html_style(...)
}
\description{
Returns a list with one element, the html style tag, whose name is the
class name given to this style.
}
\keyword{internal}
