% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{cli_tag}
\alias{cli_tag}
\title{Create a cli-formatted badge}
\usage{
cli_tag(..., scope = NULL, color = "blue", .envir = parent.frame())
}
\arguments{
\item{style}{A \code{list} of arguments to pass to the ellipsis arguments of
\code{\link[cli:make_ansi_style]{cli::make_ansi_style()}}.}
}
\description{
Create a cli-formatted badge
}
