% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data_info.R
\name{data_info}
\alias{data_info}
\title{Package Data and Metadata}
\usage{
data_info(
  metric = logical(0),
  title = character(0),
  description = structure(list(), class = "Rd"),
  tags = structure(c("adoption", "lifecycle management", "best practice", "execution",
    "transient", "version-independent"), class = c("val.meter::tags", "character",
    "S7_object"), S7_class = structure(function (...) 
 {
     n <- ...length()
    
    value <- if (n == 0L) 
         enum
     else if (n == 1L && identical(..1, TRUE)) 

            enum
     else if (n == 1L && identical(..1, FALSE)) 
         character(0L)

        else as.character(c(...))
     S7::new_object(value)
 }, name = "tags", parent =
    structure(list(class = "character", constructor_name = "character", constructor =
    function (.data = character(0)) 
 .data, validator = function (object) 
 {
     if
    (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"), package = "val.meter", properties = list(), abstract =
    FALSE, constructor = function (...) 
 {
     n <- ...length()
     value <- if (n ==
    0L) 
         enum
     else if (n == 1L && identical(..1, TRUE)) 
         enum
    
    else if (n == 1L && identical(..1, FALSE)) 
         character(0L)
     else
    as.character(c(...))
     S7::new_object(value)
 }, validator = function (self) 
 {
 
       if (!all(self \%in\% enum)) 
         fmt(error)
 }, class = c("S7_class",
    "S7_object"))),
  data_class = NULL,
  suggests = suggests(),
  permissions = permissions()
)
}
\description{
Package data is any information that can be derived about a package. Some of
these data are classified as "metrics" -- data that is structured, regular
and useful for systematic decision-making about packages. Metrics come with
additional constraints which ensure that data is calculated uniformly.
}
\section{Fields}{

\describe{
\item{\code{metric}}{\code{logical(1L)} a flag indicating whether the data is a
metric -- a piece of structured, regular package metadata useful for
decision-making about package use.}

\item{\code{title}}{\code{character(1L)} a descriptive title.}

\item{\code{description}}{\code{Rd} a short description about the data.
Should provide enough context to support effective decision-making.
Can use Rd formatting for rich text.}

\item{\code{tags}}{\code{\link[=class_tags]{class_tags()}} a set of tags, used for
classifying metrics into categories.}

\item{\code{data_class}}{an object that can be converted into an \code{S7_class}
using \code{\link[S7:as_class]{S7::as_class()}}, used for enforcing a return class on data
derivations.}

\item{\code{suggests}}{\code{character(n)} a vector of suggested packages needed
for deriving a piece of data. If the package is not available, the
metric will not be derived.}

\item{\code{permissions}}{\code{\link[=permissions]{permissions()}} a vector of enumerated permissions
that must be granted before this piece of data will be derived.}
}}

\concept{data}
