% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data_info.R
\name{data_info_list}
\alias{data_info_list}
\title{A list of package data information}
\usage{
data_info_list(.data = list())
}
\description{
This class is largely superficial. It's primary purpose is the
pretty-printing of package data metadata.
}
\keyword{internal}
