% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enum_vector.R
\name{enum_vector}
\alias{enum_vector}
\title{Construct a \code{character} subclass with a restricted set of values}
\usage{
enum_vector(name, ..., enum = c(), error)
}
\description{
Construct a \code{character} subclass with a restricted set of values
}
\keyword{internal}
