% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_evaluate.R
\name{format_output}
\alias{format_output}
\title{Formatted version of \code{capture.output()}}
\usage{
format_output(
  x,
  ...,
  style = infer_format_style(),
  evaluate = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{x}{An expression to capture output from.}

\item{...}{Additional arguments unused.}

\item{style}{What style to format output as. When \code{knitr} is running,
infers the preferred format from the output document type.}

\item{evaluate}{Whether to evaluate \code{x} by first capturing output with
\link[evaluate:evaluate]{evaluate::evaluate}. When \code{FALSE}, format \code{x} as a \code{character} value
directly.}

\item{envir}{An environment in which expression \code{x} should be evaluated.}
}
\description{
Captures output and formats it according to \code{style}. Uses the \link[evaluate:evaluate]{evaluate}
package to capture output to be more resilient to other processes sinking
output, causing issues with \code{capture.output} -- notably when used with
\link[knitr:knitr]{knitr}.
}
\examples{
format_output(
  cli::cli_text(cli::col_red("hello, world!")),
  type = "message",  # cli outputs to message stream
  style = "html"
)

}
