% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dcf.R
\name{from_dcf}
\alias{from_dcf}
\title{Parse an object from a \code{DCF}-formatted string}
\usage{
from_dcf(x, to, ...)
}
\arguments{
\item{x}{\code{character(n)} \code{DCF} vector}

\item{to}{A \code{S7::S7_object} class to convert into}

\item{...}{Additional arguments unused}
}
\value{
An object of type \code{to}
}
\description{
Parse an object from a \code{DCF}-formatted string
}
