% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pkg.R
\name{get_pkg_data}
\alias{get_pkg_data}
\title{Get \code{\link{pkg}} object data}
\usage{
get_pkg_data(x, name, ..., .raise = .state$raise)
}
\arguments{
\item{x}{\code{\link{pkg}} object to derive data for}

\item{name}{\code{character(1L)} field name for the data to derive}

\item{...}{Additional arguments unused}

\item{.raise}{\code{logical(1L)} flag indicating whether errors should be raised
or captured. This flag is not intended to be set directly, it is exposed
so that recursive calls can raise lower-level errors while capturing them
at the surface.}
}
\value{
the derived data, using the method of \code{\link{pkg_data_derive}} dispatched
on field \code{name}.
}
\description{
This methods handles the error handling and propagation of deriving package
data. It is the primary interface by which a \code{\link{pkg}} object should be
deriving data. In contrast to \code{\link{pkg_data_derive}}, which is the method that
individual data implements to register it as a field, this function wraps
the execution in appropriate error handling for user presentation. This
function should not throw errors, but instead should capture errors for
communication back to the user.
}
\keyword{internal}
