% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_evaluate.R
\name{infer_format_style}
\alias{infer_format_style}
\title{Try to infer the output format for formatting functions}
\usage{
infer_format_style()
}
\description{
Try to infer the output format for formatting functions
}
\keyword{internal}
