% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rstudio.R
\name{is_rs_rpc_get_completions_call}
\alias{is_rs_rpc_get_completions_call}
\title{Check whether we're currently evaluating RStudio's completions}
\usage{
is_rs_rpc_get_completions_call(call = sys.call(1L))
}
\arguments{
\item{call}{Which call to check. Defaults to the first call in the call
stack, which will always be the originating call of RStudio's completions
calls.}
}
\description{
Check whether we're currently evaluating RStudio's completions
}
\keyword{internal}
