% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_evaluate.R
\name{knit_print.evaluate_evaluation}
\alias{knit_print.evaluate_evaluation}
\title{\code{knitr} Pretty-printing of rich ansi output captured with \code{evaluate()}}
\usage{
knit_print.evaluate_evaluation(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
\code{knitr} Pretty-printing of rich ansi output captured with \code{evaluate()}
}
