% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data_info.R
\name{metrics}
\alias{metrics}
\title{Package metric data}
\usage{
metrics(x, ..., all = FALSE)
}
\arguments{
\item{x}{Optionally, an object to retrieve metrics from. When \code{NULL} (the
default), a listing of metric metadata is returned.}

\item{...}{Additional arguments unused.}

\item{all}{If \code{TRUE}, include non-metric package data. These are often
intermediate data used in the calculation of package metrics.}
}
\value{
A \code{list} of calculated values or metadata, in the cases where an
object is or is not provided respectively.
}
\description{
Package metric data
}
\section{Metrics}{The following metrics are provided by \code{\link{val.meter}}.
 \subsection{Dependency Count}{\code{<integer>} the number of required dependencies

  

\link[val.meter:tags]{\ifelse{html}{\figure{badge-best_practice-x-flat-square-blue.svg}{options: alt='[best practice]'}}{\strong{[best practice]}}}}}
