% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_pkg}
\alias{new_pkg}
\title{Package Class}
\usage{
new_pkg(
  data = new.env(parent = emptyenv()),
  resources = list(),
  scopes = pkg_data_permissions()
)
}
\description{
Aggregates package information, surveying metadata from discovered
resources.
}
\section{Fields}{

\describe{
\item{\code{data}}{A mutable environment, used to aggregate package meatadata.
When working with a \code{\link{pkg}} object, you may prefer to use the \code{[[} and
\code{$} operators for accessing data in this environment, which will also
prompt any necessary data dependencies to be evaluated.}

\item{\code{metrics}}{Return calculated listing of all metrics}

\item{\code{resources}}{A list of resources of package data for the
particular package. The order of the list determines the precedence
of information. If information about a package could be derived from
multiple sources, the first source is prioritized.}

\item{\code{scopes}}{Permissible scopes for deriving data.}
}}

