% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{val.meter Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{val.meter} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "val.meter")

options::opt(x, default, env = "val.meter")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{policy}{\describe{
Set the default \code{val.meter} policies, specifying how package
resources will be discovered and what permissions are granted when
calculating metrics.\item{default: }{\preformatted{policy()}}
\item{option: }{val.meter.policy}
\item{envvar: }{R_VAL_METER_POLICY (evaluated if possible, raw string otherwise)}
}}

\item{tags}{\describe{
Set the default \code{val.meter} tags policy. Tags characterize the
types of information various metrics contain. For more details, see
\code{\link[=tags]{tags()}}.\item{default: }{\preformatted{tags(TRUE)}}
\item{option: }{val.meter.tags}
\item{envvar: }{R_VAL_METER_TAGS (evaluated if possible, raw string otherwise)}
}}

\item{logs}{\describe{
Whether output should be captured during the evaluation of metrics.\item{default: }{\preformatted{FALSE}}
\item{option: }{val.meter.logs}
\item{envvar: }{R_VAL_METER_LOGS (evaluated if possible, raw string otherwise)}
}}

\item{artifacts}{\describe{
Directory where artifacts will be stored. This includes installation logs,
package source code and temporary libraries used while evaluating packages.\item{default: }{\preformatted{ns_tmp_root()}}
\item{option: }{val.meter.artifacts}
\item{envvar: }{R_VAL_METER_ARTIFACTS (evaluated if possible, raw string otherwise)}
}}

\item{quiet}{\describe{
Silences console output during evaluation. This applies when pulling package
resources (such as download and installation output) and executing code
(for example, running \verb{R CMD check})\item{default: }{\preformatted{TRUE}}
\item{option: }{val.meter.quiet}
\item{envvar: }{R_VAL_METER_QUIET (evaluated if possible, raw string otherwise)}
}}

\item{lint_linters}{\describe{
Character vector of \code{{lintr}} linter function names used when computing
\code{lint_count} and \code{lint_free_fraction}. Each name must correspond
to a linter constructor exported by \code{{lintr}} (e.g.
\code{"equals_na_linter"}. Override to add, remove, or replace linters.
Note: linters requiring configuration or environment-specific ones
(e.g. \code{"backport_linter"}) are excluded from the default set;
add them manually by changing this option, possibly after configuring
them globally.\item{default: }{\preformatted{c("absolute_path_linter", "all_equal_linter", "class_equals_linter", 
    "download_file_linter", "equals_na_linter", "for_loop_index_linter", 
    "length_test_linter", "list_comparison_linter", "nonportable_path_linter", 
    "object_overwrite_linter", "package_hooks_linter", "pipe_return_linter", 
    "redundant_equals_linter", "routine_registration_linter", 
    "sample_int_linter", "seq_linter", "sprintf_linter", "strings_as_factors_linter", 
    "T_and_F_symbol_linter", "terminal_close_linter", "unused_import_linter", 
    "vector_logic_linter")}}
\item{option: }{val.meter.lint_linters}
\item{envvar: }{R_VAL_METER_LINT_LINTERS (evaluated if possible, raw string otherwise)}
}}

\item{source_control_domains}{\describe{
Recognized source control hosting domains used when inferring whether a
package has a source code repository on a recognized hosting platform.
Customize this to add additional git hosting services (e.g., self-hosted
GitLab instances or other federated git providers).\item{default: }{\preformatted{c("github.com", "gitlab.com", "bitbucket.org", "r-forge.r-project.org", 
    "codeberg.org", "sr.ht", "gitea.com", "git.sr.ht")}}
\item{option: }{val.meter.source_control_domains}
\item{envvar: }{R_VAL_METER_SOURCE_CONTROL_DOMAINS (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
