% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_params}
\alias{options_params}
\title{Options As Parameters}
\arguments{
\item{tags}{Set the default \code{val.meter} tags policy. Tags characterize the
types of information various metrics contain. For more details, see
\code{\link[=tags]{tags()}}. (Defaults to \code{tags(TRUE)}, overwritable using option 'val.meter.tags' or environment variable 'R_VAL_METER_TAGS')}

\item{quiet}{Silences console output during evaluation. This applies when pulling package
resources (such as download and installation output) and executing code
(for example, running \verb{R CMD check}) (Defaults to \code{TRUE}, overwritable using option 'val.meter.quiet' or environment variable 'R_VAL_METER_QUIET')}

\item{policy}{Set the default \code{val.meter} policies, specifying how package
resources will be discovered and what permissions are granted when
calculating metrics. (Defaults to \code{policy()}, overwritable using option 'val.meter.policy' or environment variable 'R_VAL_METER_POLICY')}

\item{bioc_site_repository}{Additional Bioconductor site repositories to include when constructing
\code{\link[=bioc_repo_resource]{bioc_repo_resource()}} or calculating Bioconductor reverse dependencies.
Can be set to e.g. an URL to an organization's internal Bioconductor repository
or a custom mirror. This value is passed to \code{BiocManager::repositories()}. (Defaults to \code{character()}, overwritable using option 'val.meter.bioc_site_repository' or environment variable 'R_VAL_METER_BIOC_SITE_REPOSITORY')}

\item{logs}{Logging directory where artifacts will be stored. Defaults to a temporary
directory. (Defaults to \code{ns_tmp_root()}, overwritable using option 'val.meter.logs' or environment variable 'R_VAL_METER_LOGS')}

\item{source_control_domains}{Recognized source control hosting domains used when inferring whether a
package has a source code repository on a recognized hosting platform.
Customize this to add additional git hosting services (e.g., self-hosted
GitLab instances or other federated git providers). (Defaults to \verb{c("github.com", "gitlab.com", "bitbucket.org", "r-forge.r-project.org", ;     "codeberg.org", "sr.ht", "gitea.com", "git.sr.ht")}, overwritable using option 'val.meter.source_control_domains' or environment variable 'R_VAL_METER_SOURCE_CONTROL_DOMAINS')}
}
\description{
Options As Parameters
}
\keyword{internal}
