% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_permissions.R
\name{permissions}
\alias{permissions}
\title{Package Data Permissions Class}
\usage{
permissions(...)
}
\arguments{
\item{...}{Values to include in enumerated vector. Multiple values will be
concatenated into a vector before instantiation. Special handling if the
first and only argument is a logical, interpreting \code{TRUE} as a vector of
all enumerated values and \code{FALSE} as a zero-length vector.}
}
\description{
A descriptor of behaviors required for deriving package data. Each policy
property is a \code{\link{logical}} flag, which may be either \code{TRUE} or \code{FALSE}.
Policies are declared for each metric, as well as when deriving metrics for
a package. When deriving metrics, only those metrics which conform to the set
policy will be derived.
}
\details{
Given a policy, only assessments which conform to all enabled flags will be
assessed.
}
\section{Accepted Values}{
\itemize{
\item \code{"write"}
\item \code{"execution"}
\item \code{"network"}
}
}

\keyword{safeguard}
