% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pkg}
\alias{pkg}
\title{Package Class}
\usage{
pkg(
  data = new.env(parent = emptyenv()),
  resources = list(),
  scopes = pkg_data_scopes()
)
}
\description{
Aggregates package information, surveying metadata from discovered
resources.
}
\section{Fields}{

\describe{
\item{\code{data}}{A mutable environment, used to aggregate package meatadata}

\item{\code{resources}}{A list of resources of package data for the
particular package. The order of the list determines the precedence
of information. If information about a package could be derived from
multiple sources, the first source is prioritized.}

\item{\code{scopes}}{Permissible scopes for deriving data.}
}}

