% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_resources.R
\name{pkg_archive_source_resource}
\alias{pkg_archive_source_resource}
\title{Package Archive Source Code Resource Class}
\usage{
pkg_archive_source_resource(path = NA_character_)
}
\description{
The extracted source code from a package's build archive.
}
