% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pkg_data_derive.R
\name{pkg_data_derive}
\alias{pkg_data_derive}
\title{Derive a package data field}
\usage{
pkg_data_derive(pkg, resource, field, ...)
}
\arguments{
\item{pkg}{A \code{\link[=pkg]{pkg()}}}

\item{resource}{A \code{\link[=resource]{resource()}}, or if not provided, the \code{\link{resource}}
extracted from \code{pkg@resource}.}

\item{field}{Used for dispatching on which field to derive. Methods are
provided such that a simple \code{character} field name can be passed and
used to build a class for dispatching to the right derivation function.}

\item{...}{Used by specific methods.}
}
\value{
The derived field value.
}
\description{
Derive a \code{\link{pkg}} data field, giving a function by which a piece of package
data is calculated. This function is not called directly, it used by the
indexing functions implemented for \code{\link{pkg}} to populate new data fields.
Default implementations accept \code{field} as a \code{character}, automatically
dispatching on the field as a class object, and accept a missing
\code{resource}, instead iterating through \code{pkg} resources by priority.
}
\details{
This function is used internally when accessing a \code{\link{pkg}} object using \code{$},
\code{[[} and \code{[}.
}
\seealso{
Other generics:
\code{\link[=pkg_data_info]{pkg_data_info()}}
}
\concept{generics}
