% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pkg_data_derive_error}
\alias{pkg_data_derive_error}
\title{Package Data Function Class}
\usage{
pkg_data_derive_error(.data)
}
\description{
A wrapper around a function, which allows binding required scopes.
}
