% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pkg_data_info.R
\name{pkg_data_info}
\alias{pkg_data_info}
\title{Derive package data information}
\usage{
pkg_data_info(field, resource, ...)
}
\arguments{
\item{field}{\code{character(1L)} field name to retrieve.}

\item{resource}{\code{\link{resource}} providing the resource to be used for
retrieving package data info.}

\item{...}{Additional arguments unused.}
}
\value{
\code{data_info} for field \code{field}.
}
\description{
Retrieve metadata about a data field.
}
\seealso{
Other generics:
\code{\link[=pkg_data_derive]{pkg_data_derive()}}
}
\concept{generics}
