% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pkg_data_permissions}
\alias{pkg_data_permissions}
\title{Package Data Scopes Class}
\usage{
pkg_data_permissions(...)
}
\description{
A descriptor of behaviors required for deriving package data. Each policy
property is a \code{\link{logical}} flag, which may be either \code{TRUE} or \code{FALSE}.
Policies are declared for each metric, as well as when deriving metrics for
a package. When deriving metrics, only those metrics which conform to the set
policy will be derived.
}
\details{
Given a policy, only assessments which conform to all enabled flags will be
assessed.
}
