% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{pkg_data_tags}
\alias{pkg_data_tags}
\title{Package Data Tags Class}
\usage{
pkg_data_tags(.data = character(0))
}
\description{
A enumeration of tags that can be applied to a metric, to capture metadata
more effectively.
}
