% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pkg.R
\name{pkg_from_dcf}
\alias{pkg_from_dcf}
\alias{pkgs_from_dcf}
\title{Produce \code{pkg}(s) from a \code{DCF}-formatted string}
\usage{
pkg_from_dcf(x, ...)

pkgs_from_dcf(x, ...)
}
\arguments{
\item{x}{A \code{DCF}-formatted string}

\item{...}{Additional arguments unused}
}
\value{
A \code{pkg} object. Note that when parsing from a \code{DCF} string.
}
\description{
Produce \code{pkg}(s) from a \code{DCF}-formatted string
}
