% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_resources.R
\name{pkg_git_resource}
\alias{pkg_git_resource}
\title{Package \code{git} Resource Class}
\usage{
pkg_git_resource(http_url = character(0))
}
\description{
A reference to a listing in an R package git source code repository.
}
