% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_resources.R
\name{pkg_install_resource}
\alias{pkg_install_resource}
\title{Package Install Resource Class}
\usage{
pkg_install_resource(path = NA_character_)
}
\description{
An installed version of a package, as would be found in a package library.
}
