% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_resources.R
\docType{data}
\name{pkg_local_resource}
\alias{pkg_local_resource}
\title{Package Local Resource Class Union}
\format{
An object of class \code{S7_union} of length 1.
}
\usage{
pkg_local_resource
}
\description{
A union of all package resource classes that have local source code.
}
\keyword{datasets}
