% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_metrics.R
\name{pkg_metric}
\alias{pkg_metric}
\title{Package Metric Classes}
\description{
Package metrics are like any other package data, but have additional
metadata and are constrained to only atomic types.
}
