% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_resources.R
\name{pkg_repo_resource}
\alias{pkg_repo_resource}
\title{Package Repository Resource Class}
\usage{
pkg_repo_resource(
  package = character(0),
  version = character(0),
  repo = character(0)
)
}
\description{
A reference to a listing in an R package repository.
}
