% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_resources.R
\name{pkg_resource}
\alias{pkg_resource}
\title{Package Resource Class}
\usage{
pkg_resource()
}
\description{
A package resource is a resource for producing package information. Package
resources can vary from source code repositories to R package repository
listings. Each resources should be able to produce a downloadable version
of executable package code, but the extent of included code might vary
depending on source.
}
\details{
Package resources should also implement \code{convert()}, providing a method for
converting into a more desirable source of information. For example, given
only a listing of a package in a repository, calling \code{convert()} on such
a reference may populate a local directory with the source code of the
package.
}
