% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{pkg_words}
\alias{pkg_words}
\title{Words for Package Names}
\format{
A data frame with 1394 rows and 3 variables:
\describe{
\item{word}{words with positive sentiment}
\item{pos}{part of speech}
\item{weight}{an entirely non-scientific weighting to use when sampling
words to use for package names. Gives extra weight to some science and
statistical words and R-isms that would make for fun surprises.}
}
}
\source{
tidytext
}
\usage{
pkg_words
}
\description{
A dataset containing positive words and their associated parts of speech.
Words are selected from the \code{tidytext} \code{parts_of_speech} dataset, filtered
for words with positive associations.
}
\keyword{datasets}
