% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_policy.R
\name{policy}
\alias{policy}
\title{Package Evaluation Policy Class}
\usage{
policy(
  accepted_resources = list(
    source_archive_resource,
    source_code_resource,
    install_resource
  ),
  source_resources = list(
    cran_repo_resource,
    repo_resource
  ),
  permissions = class_permissions(FALSE)
)
}
\arguments{
\item{accepted_resources}{A list of resources types to permit. Ordered by
priority, highest to lowest.}

\item{source_resources}{A list of additional resource types, which may be
used to discover a resource of an accepted type. For example, even if
only \code{\link[=source_archive_resource]{source_archive_resource()}}s are accepted, a \code{\link[=repo_resource]{repo_resource()}}
could be used as a means of acquiring the archive source, so long as it
can be \code{\link[S7:convert]{S7::convert()}}ed into a \code{\link[=source_archive_resource]{source_archive_resource()}}.}

\item{permissions}{Behavioral permissions provided for resource
acquisition. For example, downloading and installing source code for
more accurate metric evaluation requires the \code{"network"} and \code{"write"}
permissions.}
}
\description{
A descriptor of how package resources should be discovered, indicating
which types of package resources should be considered and how they must be
used to produce consistently sourced information.
}
\details{
The policy takes effect when packages are passed to \code{\link[=pkg]{pkg()}}, limiting how
package resources can be discovered. A policy can be applied globally using
provided \code{\link{options}}.
}
\examples{
\dontrun{
# discover locally installed file path, create `pkg` from `local_resource`
pkg(find.package("val.meter"))

# disable local resource discovery
options(val.meter.policy = policy(
  accepted_resources = list(source_archive_resource)
))

# expect error - unable to discover resource
tryCatch(
  pkg(find.package("val.meter")),
  error = function(error, ...) message(error$message)
)
}

}
\keyword{safeguard}
