% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pkg.R
\name{random-pkg}
\alias{random-pkg}
\alias{random_pkg}
\alias{random_pkgs}
\title{Generate Random Package(s)}
\usage{
random_pkg(package = random_pkg_name(), version = random_pkg_version(), ...)

random_pkgs(n = 100, ...)
}
\arguments{
\item{package}{\code{character(1L)} a package name}

\item{version}{\code{character(1L)} a package version}

\item{...}{Additional arguments passed to \code{pkg}}

\item{n}{\code{integer(1L)} how many packages to simulate}
}
\description{
Create a package object to simulate metric derivation. When generating a
collection of packages, dependencies will realistically be made between
packages.
}
