% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enum_vector.R
\name{roxygenize_enum_vector}
\alias{roxygenize_enum_vector}
\title{Roxygenize a class object derived with \code{enum_vector}}
\usage{
roxygenize_enum_vector(class)
}
\description{
Roxygenize a class object derived with \code{enum_vector}
}
\keyword{internal}
