% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{source_resource}
\alias{source_resource}
\title{Package Source Code Resource Class}
\usage{
source_resource(path = NA_character_)
}
\description{
Note that in some cases, this is distinct from the source code extracted
from a package archive, as some files are sometimes excluded from package
builds (uploaded to repositories). This can include additional development-
related metadata files, but can sometimes include additional tests that take
too long to run by automated systems by the repository host.
}
\details{
Commonly, this may mean omitting tests to avoid failures on CRAN's automated
build systems.
}
