% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dcf.R
\name{to_dcf}
\alias{to_dcf}
\title{Convert an object into a \code{DCF}-formatted string}
\usage{
to_dcf(x, ...)
}
\arguments{
\item{x}{An object to convert}

\item{...}{Additional arguments unused}
}
\value{
A \code{DCF}-formatted string
}
\description{
Convert an object into a \code{DCF}-formatted string
}
