% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rd.R
\name{utils-rd}
\alias{utils-rd}
\alias{rd_empty}
\alias{rd_escape}
\alias{rd_to_txt}
\alias{rd_parse}
\alias{rd_deparse}
\alias{rd_sexpr}
\alias{rd_badge}
\title{Rd utilities}
\usage{
rd_empty()

rd_escape(text)

rd_to_txt(...)

rd_parse(text, fragment = FALSE, permissive = TRUE, ...)

rd_deparse(rd, deparse = TRUE, ...)

rd_sexpr(
  code,
  stage = c("build", "install", "render"),
  results = c("text", "verbatim", "rd", "hide"),
  quote = TRUE
)

rd_badge(
  message,
  label = "",
  style = "",
  dest = NULL,
  color = "blue",
  url = "https://img.shields.io/badge/",
  params = list(style = "flat-square")
)
}
\description{
These functions are used primarily for generating documentation
programmatically - most notably for documenting metrics that are already
documented using \code{\link{pkg_data_info}} metadata objects.
}
\section{Functions}{
\itemize{
\item \code{rd_empty()}: Create an empty Rd object

\item \code{rd_escape()}: Escape an Rd string

\item \code{rd_to_txt()}: Convert Rd to text-formatted character vector

\item \code{rd_parse()}: Parse text into an Rd format

\item \code{rd_deparse()}: Deparse an Rd object into a character value

\item \code{rd_sexpr()}: Convert R code to an Rd \\Sexpr

\item \code{rd_badge()}: Generate a badge, using shields.io and caching svg images for display in
html output.

}}
\keyword{internal}
