% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metric_data_frame.R
\name{class_metric_data_frame}
\alias{class_metric_data_frame}
\title{Metric data frame}
\usage{
class_metric_data_frame(.data = list(), row.names = NULL)
}
\description{
A \code{\link{data.frame}}, for all intents and purposes, with column names
corresponding to \code{\link{pkg}} metric data.
}
