% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_err.R
\name{error}
\alias{error}
\title{Build an error from an error type and data attributes}
\usage{
error(type, ...)
}
\description{
Build an error from an error type and data attributes
}
\note{
This function is only intended for use when parsing serialized output
from text.

This function uses an error class to build an error object. It intentionally
produces incomplete error objects, lacking the error backtrace
call.

For signalling errors internal to the package, see \code{\link[=err]{err()}}. This function
is used to provide a readable syntax to exported \code{PACKAGES} files, which
are parsed using this function back into their respective error objects.
}
\examples{
# given a DCF input such as

## Package: testpkg
## Version: 1.2.3
## Metric/word_count@R: error("missing_suggests", "wordcount")

# we want to parse (by evaluation) the output into our own error type
error("missing_suggests", "wordcount")

}
