% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{git_resource}
\alias{git_resource}
\title{Package \code{git} Resource Class}
\usage{
git_resource(
  package = NA_character_,
  version = NA_character_,
  id = next_id(),
  md5 = NA_character_,
  http_url = character(0)
)
}
\arguments{
\item{http_url}{The git repository url}
}
\description{
A reference to a listing in an R package git source code repository.
}
\seealso{
Other resources: 
\code{\link{cran_repo_resource}()},
\code{\link{install_resource}()},
\code{\link{local_resource}()},
\code{\link{local_source_resource}()},
\code{\link{mock_resource}()},
\code{\link{multi_resource}()},
\code{\link{remote_resource}()},
\code{\link{repo_resource}()},
\code{\link{resource}()},
\code{\link{source_archive_resource}()},
\code{\link{source_code_resource}()},
\code{\link{unknown_resource}()}
}
\concept{resources}
