% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{git_resource}
\alias{git_resource}
\title{Package \code{git} Resource Class}
\usage{
git_resource(http_url = character(0))
}
\description{
A reference to a listing in an R package git source code repository.
}
