% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{http_resource}
\alias{http_resource}
\title{Package \code{http(s)} Archive Resource Class}
\usage{
http_resource(
  package = NA_character_,
  version = NA_character_,
  id = next_id(),
  md5 = NA_character_,
  http_url = character(0)
)
}
\arguments{
\item{package}{\code{character(1L)} Package name. Optional, but should be
provided if possible.}

\item{version}{\code{character(1L)} Package version, provided as a string.}

\item{id}{\code{integer(1L)} optional id used for tracking resources
throughout execution. Generally not provided directly, as new objects
automatically get a unique identifier. For example, the package source
code from a \code{\link[=repo_resource]{repo_resource()}} may be downloaded to add a
\code{\link[=source_archive_resource]{source_archive_resource()}} and add it to a new \code{\link[=multi_resource]{multi_resource()}}.
Because all of these represent the same package, they retain the same
\code{id}. Primarily the \code{id} is used for isolating temporary files.}

\item{md5}{\code{character(1L)} md5 digest of the package source code tarball.
This is not generally provided directly, but is instead derived when
acquiring resources.}

\item{http_url}{The git repository url}
}
\description{
A web-based reference to an R package file. Most often this will be used
to resolve an http url into a source code archive.
}
\seealso{
Other resources:
\code{\link[=cran_repo_resource]{cran_repo_resource()}},
\code{\link[=git_resource]{git_resource()}},
\code{\link[=install_resource]{install_resource()}},
\code{\link[=local_resource]{local_resource()}},
\code{\link[=local_source_resource]{local_source_resource()}},
\code{\link[=mock_resource]{mock_resource()}},
\code{\link[=multi_resource]{multi_resource()}},
\code{\link[=remote_resource]{remote_resource()}},
\code{\link[=repo_resource]{repo_resource()}},
\code{\link[=resource]{resource()}},
\code{\link[=source_archive_resource]{source_archive_resource()}},
\code{\link[=source_code_resource]{source_code_resource()}},
\code{\link[=unknown_resource]{unknown_resource()}}
}
\concept{resources}
