% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{install_resource}
\alias{install_resource}
\title{Package Install Resource Class}
\usage{
install_resource(path = NA_character_)
}
\description{
An installed version of a package, as would be found in a package library.
}
