% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\docType{data}
\name{local_resource}
\alias{local_resource}
\title{Package Local Resource Class Union}
\format{
An object of class \code{S7_union} of length 1.
}
\usage{
local_resource
}
\description{
A union of all package resource classes that have local source code.
}
\keyword{datasets}
