% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data_info.R
\name{metrics}
\alias{metrics}
\title{Catalog or calculate package metrics}
\usage{
metrics(x, ..., all = FALSE)
}
\arguments{
\item{x}{Optionally, an object to retrieve metrics from. When \code{NULL} (the
default), a listing of metric metadata is returned.}

\item{...}{Additional arguments unused.}

\item{all}{If \code{TRUE}, include non-metric package data. These are often
intermediate data used in the calculation of package metrics.}
}
\value{
A \code{list} of calculated values or metadata, in the cases where an
object is or is not provided respectively.
}
\description{
When no object is passed, returns a list of possible metrics. When a \code{\link{pkg}}
object is provided, return metrics calculated for that package. Metrics are
a subset of all the data calculated over the course of assessing a package.
For access to \emph{all} the internally calculated data, pass \code{all = TRUE}.
}
\keyword{workflow}
\section{Metrics}{The following metrics are provided by \code{\link{val.meter}}.
 \subsection{Total Vignettes}{
\code{<integer>} total number of vignettes

  


}
 \subsection{R CMD check Error Count}{
\code{<integer>} the number of errors produced when running \code{R CMD check}

  

\Sexpr[stage=install,results=rd]{if (numeric_version(paste0(R.version$major, ".", R.version$minor)) < "4.5.0") {     "\\\\ifelse{html}{\\\\figure{badge-execution-x-flat-square-blue.svg}{options: alt = \\"[execution]\\"}}{\\\\strong{[execution]}}" } else {     "\\\\link[val.meter:tags]{\\\\ifelse{html}{\\\\figure{badge-execution-x-flat-square-blue.svg}{options: alt = \\"[execution]\\"}}{\\\\strong{[execution]}}}" }}
}
 \subsection{Total Downloads}{
\code{<integer>} total number of lifetime downloads, as reported by the Posit CRAN mirror through the \href{https://cranlogs.rpkg.org}{cranlogs} API

\Sexpr[stage=render,results=rd]{if (!is.na(match("network", getOption("val.meter.permissions")))) "\\\\ifelse{html}{\\\\figure{badge-req-network-x-flat-square-green.svg}{options: alt = \\"[network]\\"}}{\\\\strong{[req::network]}}" else "\\\\ifelse{html}{\\\\figure{badge-req-network-x-flat-square-red.svg}{options: alt = \\"[network]\\"}}{\\\\strong{[req::network]}}"}  

\Sexpr[stage=install,results=rd]{if (numeric_version(paste0(R.version$major, ".", R.version$minor)) < "4.5.0") {     "\\\\ifelse{html}{\\\\figure{badge-adoption-x-flat-square-blue.svg}{options: alt = \\"[adoption]\\"}}{\\\\strong{[adoption]}}\\n\\\\ifelse{html}{\\\\figure{badge-transient-x-flat-square-blue.svg}{options: alt = \\"[transient]\\"}}{\\\\strong{[transient]}}\\n\\\\ifelse{html}{\\\\figure{badge-version--independent-x-flat-square-blue.svg}{options: alt = \\"[version-independent]\\"}}{\\\\strong{[version-independent]}}" } else {     "\\\\link[val.meter:tags]{\\\\ifelse{html}{\\\\figure{badge-adoption-x-flat-square-blue.svg}{options: alt = \\"[adoption]\\"}}{\\\\strong{[adoption]}}}\\n\\\\link[val.meter:tags]{\\\\ifelse{html}{\\\\figure{badge-transient-x-flat-square-blue.svg}{options: alt = \\"[transient]\\"}}{\\\\strong{[transient]}}}\\n\\\\link[val.meter:tags]{\\\\ifelse{html}{\\\\figure{badge-version--independent-x-flat-square-blue.svg}{options: alt = \\"[version-independent]\\"}}{\\\\strong{[version-independent]}}}" }}
}
 \subsection{Dependency Count}{
\code{<integer>} the number of required dependencies

  

\Sexpr[stage=install,results=rd]{if (numeric_version(paste0(R.version$major, ".", R.version$minor)) < "4.5.0") {     "\\\\ifelse{html}{\\\\figure{badge-best_practice-x-flat-square-blue.svg}{options: alt = \\"[best practice]\\"}}{\\\\strong{[best practice]}}" } else {     "\\\\link[val.meter:tags]{\\\\ifelse{html}{\\\\figure{badge-best_practice-x-flat-square-blue.svg}{options: alt = \\"[best practice]\\"}}{\\\\strong{[best practice]}}}" }}
}
 \subsection{Has Website}{
\code{<logical>} a logical indicating whether the package has a website

  


}}
