% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data_info.R
\name{metrics}
\alias{metrics}
\title{Catalog or calculate package metrics}
\usage{
metrics(x, ..., all = FALSE)
}
\arguments{
\item{x}{Optionally, an object to retrieve metrics from. When \code{NULL} (the
default), a listing of metric metadata is returned.}

\item{...}{Additional arguments unused.}

\item{all}{If \code{TRUE}, include non-metric package data. These are often
intermediate data used in the calculation of package metrics.}
}
\value{
A \code{list} of calculated values or metadata, in the cases where an
object is or is not provided respectively.
}
\description{
When no object is passed, returns a list of possible metrics. When a \code{\link{pkg}}
object is provided, return metrics calculated for that package. Metrics are
a subset of all the data calculated over the course of assessing a package.
For access to \emph{all} the internally calculated data, pass \code{all = TRUE}.
}
\keyword{workflow}
\section{Metrics}{The following metrics are provided by \code{\link{val.meter}}.
 \subsection{Unit tests expression coverage}{
\code{<double>} the fraction of expressions evaluated by any unit tests

  


}
 \subsection{Dependency Count}{
\code{<integer>} the number of required dependencies

  

\Sexpr[stage=install,results=rd]{if (numeric_version(paste0(R.version$major, ".", R.version$minor)) < "4.5.0") {     "\\\\ifelse{html}{\\\\figure{badge-best_practice-x-flat-square-blue.svg}{options: alt = \\"[best practice]\\"}}{\\\\strong{[best practice]}}" } else {     "\\\\link[val.meter:tags]{\\\\ifelse{html}{\\\\figure{badge-best_practice-x-flat-square-blue.svg}{options: alt = \\"[best practice]\\"}}{\\\\strong{[best practice]}}}" }}
}}
