% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{mock_resource}
\alias{mock_resource}
\title{Mocked Package Resource}
\usage{
mock_resource(
  package = NA_character_,
  version = NA_character_,
  id = next_id(),
  md5 = NA_character_
)
}
\arguments{
\item{package}{\code{character(1L)} Package name. Optional, but should be
provided if possible.}

\item{version}{\code{character(1L)} Package version, provided as a string.}

\item{id}{\code{integer(1L)} optional id used for tracking resources
throughout execution. Generally not provided directly, as new objects
automatically get a unique identifier. For example, the package source
code from a \code{\link[=repo_resource]{repo_resource()}} may be downloaded to add a
\code{\link[=archive_source_resource]{archive_source_resource()}} and add it to a new \code{\link[=multi_resource]{multi_resource()}}.
Because all of these represent the same package, they retain the same
\code{id}. Primarily the \code{id} is used for isolating temporary files.}

\item{md5}{\code{character(1L)} md5 digest of the package source code tarball.
This is not generally provided directly, but is instead derived when
acquiring resources.}
}
\description{
The mocked resource is not intended to be used for deriving real package
data. Its purpose is as a signal to internals that we want to generate
fake, or "mocked", package data. The only practical place where a mocked
resource is needed is when adding your own data implementations and adding
a custom data simulation method.
}
\seealso{
Other resources: 
\code{\link{cran_repo_resource}()},
\code{\link{git_resource}()},
\code{\link{install_resource}()},
\code{\link{local_resource}()},
\code{\link{local_source_resource}()},
\code{\link{multi_resource}()},
\code{\link{remote_source_resource}()},
\code{\link{repo_resource}()},
\code{\link{resource}()},
\code{\link{source_archive_resource}()},
\code{\link{source_code_resource}()},
\code{\link{unknown_resource}()}
}
\concept{resources}
