% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{val.meter Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{val.meter} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "val.meter")

options::opt(x, default, env = "val.meter")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{policy}{\describe{
Set the default \code{val.meter} policies, specifying how package
resources will be discovered and what permissions are granted when
calculating metrics.\item{default: }{\preformatted{policy()}}
\item{option: }{val.meter.policy}
\item{envvar: }{R_VAL_METER_POLICY (evaluated if possible, raw string otherwise)}
}}

\item{tags}{\describe{
Set the default \code{val.meter} tags policy. Tags characterize the
types of information various metrics contain. For more details, see
\code{\link[=tags]{tags()}}.\item{default: }{\preformatted{tags(TRUE)}}
\item{option: }{val.meter.tags}
\item{envvar: }{R_VAL_METER_TAGS (evaluated if possible, raw string otherwise)}
}}

\item{logs}{\describe{
Logging directory where artifacts will be stored. Defaults to a temporary
directory.\item{default: }{\preformatted{ns_tmp_root()}}
\item{option: }{val.meter.logs}
\item{envvar: }{R_VAL_METER_LOGS (evaluated if possible, raw string otherwise)}
}}

\item{quiet}{\describe{
Silences console output during evaluation. This applies when pulling package
resources (such as download and installation output) and executing code
(for example, running \verb{R CMD check})\item{default: }{\preformatted{TRUE}}
\item{option: }{val.meter.quiet}
\item{envvar: }{R_VAL_METER_QUIET (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
