% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{val.meter Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{val.meter} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "val.meter")

options::opt(x, default, env = "val.meter")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{policy}{\describe{
Set the default val.meter package search policy.\item{default: }{\preformatted{pkg_resource_policy()}}
\item{option: }{val.meter.policy}
\item{envvar: }{R_VAL_METER_POLICY (evaluated if possible, raw string otherwise)}
}}

\item{scopes}{\describe{
Set the default val.meter data scopes policy. Scopes specify
permissions permitted when deriving package data. For more details, see
\code{?pkg_data_scopes}.\item{default: }{\preformatted{pkg_data_scopes()}}
\item{option: }{val.meter.scopes}
\item{envvar: }{R_VAL_METER_SCOPES (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
