% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_params}
\alias{options_params}
\title{Options As Parameters}
\arguments{
\item{policy}{Set the default val.meter package search policy. (Defaults to \code{pkg_resource_policy()}, overwritable using option 'val.meter.policy' or environment variable 'R_VAL_METER_POLICY')}

\item{scopes}{Set the default val.meter data scopes policy. Scopes specify
permissions permitted when deriving package data. For more details, see
\code{?pkg_data_scopes}. (Defaults to \code{pkg_data_scopes()}, overwritable using option 'val.meter.scopes' or environment variable 'R_VAL_METER_SCOPES')}
}
\description{
Options As Parameters
}
