% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_params}
\alias{options_params}
\title{Options As Parameters}
\arguments{
\item{tags}{Set the default val.meter tags policy. Tags characterize the types
of information various metrics contain. For more details, see
\code{\link[=pkg_data_tags]{pkg_data_tags()}}. (Defaults to \code{pkg_data_tags(TRUE)}, overwritable using option 'val.meter.tags' or environment variable 'R_VAL_METER_TAGS')}

\item{policy}{Set the default val.meter package search policy. (Defaults to \code{pkg_resource_policy()}, overwritable using option 'val.meter.policy' or environment variable 'R_VAL_METER_POLICY')}

\item{permissions}{Set the default val.meter data permissions policy. Permissions
specify what capabilities are permitted when deriving package data. For more
details, see \code{\link[=pkg_data_permissions]{pkg_data_permissions()}}. (Defaults to \code{pkg_data_permissions(FALSE)}, overwritable using option 'val.meter.permissions' or environment variable 'R_VAL_METER_PERMISSIONS')}
}
\description{
Options As Parameters
}
