% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_params}
\alias{options_params}
\title{Options As Parameters}
\arguments{
\item{tags}{Set the default \code{val.meter} tags policy. Tags characterize the
types of information various metrics contain. For more details, see
\code{\link[=tags]{tags()}}. (Defaults to \code{tags(TRUE)}, overwritable using option 'val.meter.tags' or environment variable 'R_VAL_METER_TAGS')}

\item{artifacts}{Directory where artifacts will be stored. This includes installation logs,
package source code and temporary libraries used while evaluating packages. (Defaults to \code{ns_tmp_root()}, overwritable using option 'val.meter.artifacts' or environment variable 'R_VAL_METER_ARTIFACTS')}

\item{quiet}{Silences console output during evaluation. This applies when pulling package
resources (such as download and installation output) and executing code
(for example, running \verb{R CMD check}) (Defaults to \code{TRUE}, overwritable using option 'val.meter.quiet' or environment variable 'R_VAL_METER_QUIET')}

\item{policy}{Set the default \code{val.meter} policies, specifying how package
resources will be discovered and what permissions are granted when
calculating metrics. (Defaults to \code{policy()}, overwritable using option 'val.meter.policy' or environment variable 'R_VAL_METER_POLICY')}

\item{logs}{Whether output should be captured during the evaluation of metrics. (Defaults to \code{TRUE}, overwritable using option 'val.meter.logs' or environment variable 'R_VAL_METER_LOGS')}
}
\description{
Options As Parameters
}
\keyword{internal}
