% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_pkg_data.R
\name{pkg_data}
\alias{pkg_data}
\alias{pkg_data_derive}
\alias{pkg_data_info}
\alias{impl_data}
\alias{impl_data_info}
\alias{impl_data_derive}
\title{Package Data Trait}
\usage{
pkg_data_derive(field, pkg, resource, ...)

pkg_data_info(field, ...)

impl_data(name, fn, for_resource = resource, ..., metric = FALSE)

impl_data_info(
  name,
  class = class_any,
  description = "",
  tags = class_tags(c()),
  permissions = class_permissions(c()),
  suggests = character(0L),
  metric = FALSE
)

impl_data_derive(name, fn, resource)
}
\description{
A trait is a collection of methods. Implemented a new piece of package data
requires that all of the following methods are implemented or have a provided
default implementation.
}
\details{
Although it's possibly to implement all of these methods one by one, it will
be easier to use \code{\link[=impl_data]{impl_data()}}, which takes only the expected data and
uses it to implemented each method.
}
\section{Functions}{
\itemize{
\item \code{pkg_data_derive()}: Derive a \code{\link{pkg}} data field, giving a function by which a piece of package
data is calculated. This function is not called directly, it used by the
indexing functions implemented for \code{\link{pkg}} to populate new data fields.
Default implementations accept \code{field} as a \code{character}, automatically
dispatching on the field as a class object, and accept a missing
\code{resource}, instead iterating through \code{pkg} resources by priority.

\item \code{pkg_data_info()}: Retrieve metadata about a data field.

\item \code{impl_data()}: Helper for implementing all the necessary methods for package data.
Internally this is a wrapper for \verb{[impl_data_meta()]} (associate metadata
with the data field), \verb{[impl_data_derive()]} (associate a derivation
function for a combination of data field \emph{and} package resource) and
\verb{[impl_metric()]} (declare a piece of data to be a metric).

\item \code{impl_data_info()}: Associate metadata with the data field

\item \code{impl_data_derive()}: Register a derivation function for a data field and package resource.

}}
