% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pkg_data_derive}
\alias{pkg_data_derive}
\title{Derive Package Data}
\usage{
pkg_data_derive(field, pkg, resource, ...)
}
\description{
Package datacan be any relevant bit of data pertaining to a package.
This can be intrinsic qualities of the package like data read directly from
a description file or the number of vignette files bundled with the package
itself. It can also be derived from executing the package code, such as
unit test results, or pulled from external sources such as the number of
lifetime package downloads.
}
\examples{
p <- pkg_from("glue")
derive("desc", p)

}
