% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{repo_resource}
\alias{repo_resource}
\title{Package Repository Resource Class}
\usage{
repo_resource(
  package = NA_character_,
  version = NA_character_,
  id = next_id(),
  md5 = NA_character_,
  repo = character(0)
)
}
\description{
A reference to a listing in an R package repository.
}
