% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{repo_resource}
\alias{repo_resource}
\title{Package Repository Resource Class}
\usage{
repo_resource(
  package = character(0),
  version = character(0),
  repo = character(0)
)
}
\description{
A reference to a listing in an R package repository.
}
