% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_resource.R
\name{source_code_resource}
\alias{source_code_resource}
\title{Package Local Resource Class Union}
\usage{
source_code_resource(
  id = next_id(),
  md5 = NA_character_,
  package = NA_character_,
  version = NA_character_,
  path = NA_character_
)
}
\description{
A union of all package resource classes that have local source code.
}
