% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{val.meter-package}
\alias{val.meter}
\alias{val.meter-package}
\title{val.meter: Measures packages for common best practices, adoption, and code execution}
\description{
Measures packages for common best practices, adoption and code execution. Built with regulated industries in mind, supporting the surveying of packages and aggregating findings into a portable format to support activities like selecting mature packages and reporting on risk-based software decision-making.
}
\author{
\strong{Maintainer}: Doug Kelkhoff \email{doug.kelkhoff@gmail.com}

Other contributors:
\itemize{
  \item R Validation Hub \email{r-val-hub-maintainers@lists.r-consortium.org} (https://www.pharmar.org) [copyright holder]
}

}
\keyword{internal}
